requirements.txt
Django==1.11.9
django-filter==1.1.0
djangorestframework==3.7.7
Pillow==5.0.0
psycopg2==2.7.3.2
social-auth-app-django==2.1.0
django-rest-framework-social-oauth2==1.1.0